requirements.txt
# Core FastAPI dependencies
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
pydantic

# Data processing
pandas
yfinance
riskfolio-lib
quantstats

# Airflow with constraints to avoid conflicts
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-3.0.4/constraints-3.11.txt"
apache-airflow==3.0.4
apache-airflow-providers-fab
flask_appbuilder
